rename 'utility' to 'utilities' for consistency
