+++ src/main/java/com/example/pupsis_main_dashboard/PUPSIS.java
@@ -6,10 +6,8 @@
 package com.example.pupsis_main_dashboard;
 
 import com.example.pupsis_main_dashboard.controllers.SettingsController; 
-import com.example.pupsis_main_dashboard.utility.StageAndSceneUtils;
+import com.example.pupsis_main_dashboard.utilities.StageAndSceneUtils;
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 import javafx.stage.Window;
