+++ src/main/java/com/example/pupsis_main_dashboard/utilities/Student.java
@@ -1,4 +1,4 @@
-package com.example.pupsis_main_dashboard.utility;
+package com.example.pupsis_main_dashboard.utilities;
 
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
